settings to its own component
